functions/requirements.txt
firebase_functions~=0.1.0
httpx
python-dotenv